rename nipoppy.tabular.bagel module
