+++ nipoppy/tabular/__init__.py
@@ -1,5 +1,5 @@
 """Classes for tabular data representation/manipulation."""
 
-from .bagel import ProcessingStatus
 from .doughnut import Doughnut, generate_doughnut, update_doughnut
 from .manifest import Manifest
+from .processing_status import ProcessingStatus
