+++ nipoppy/workflows/base.py
@@ -18,11 +18,11 @@ from nipoppy.config.main import Config
 from nipoppy.env import PROGRAM_NAME, ReturnCode, StrOrPathLike
 from nipoppy.layout import DatasetLayout
 from nipoppy.logger import add_logfile, capture_warnings, get_logger
-from nipoppy.tabular.bagel import ProcessingStatus
 from nipoppy.tabular.base import BaseTabular
 from nipoppy.tabular.dicom_dir_map import DicomDirMap
 from nipoppy.tabular.doughnut import Doughnut, generate_doughnut
 from nipoppy.tabular.manifest import Manifest
+from nipoppy.tabular.processing_status import ProcessingStatus
 from nipoppy.utils import add_path_timestamp, process_template_str
 
 LOG_SUFFIX = ".log"
